refactor: rename handler struct and methods for consistency; remove unused person package
